Update snapshot model to spring entity
