--- src/main/java/org/fourhorde/coolmemes/model/Snapshot.java
@@ -1,14 +1,0 @@
-package org.fourhorde.coolmemes.model;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-
-@Entity
-public class Snapshot {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-
-}
